cleanup bad files/changes

+++ src/java/boa/evaluator/BoaEvaluator.java
@@ -191,17 +191,15 @@ public class BoaEvaluator extends BoaMain {
 			final File outputDir = new File(this.OUTPUT_DIR);
 
 			if (outputDir.exists()) {
-				FileUtils.deleteDirectory(outputDir);
-				
-//				System.err.print("output directory '" + this.OUTPUT_DIR + "' exists - delete? [Y/n] ");
-//				int b = (char)System.in.read();
-//				char ch = (char)b;
-//
-//				if (b == -1 || ch == 'Y' || ch == 'y' || ch == '\n' || ch == '\r') {
-//					FileUtils.deleteDirectory(outputDir);
-//				} else {
-//					throw new RuntimeException("Please remove or provide a different output directory.");
-//				}
+				System.err.print("output directory '" + this.OUTPUT_DIR + "' exists - delete? [Y/n] ");
+				int b = (char)System.in.read();
+				char ch = (char)b;
+
+				if (b == -1 || ch == 'Y' || ch == 'y' || ch == '\n' || ch == '\r') {
+					FileUtils.deleteDirectory(outputDir);
+				} else {
+					throw new RuntimeException("Please remove or provide a different output directory.");
+				}
 			}
 		}
